backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic[dotenv]
passlib[async]
python-jose
python-dotenv
pyjwt
pydantic[email]
python-multipart
sqlalchemy
asyncpg 
sqlalchemy[asyncio]
asyncpg
dnspython
quixstreams
aiosmtplib
bcrypt